pyproject.toml
[project]
name = "gmail-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi>=0.116.1",
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.3",
    "uvicorn>=0.35.0",
]
